prop naming and url params added

+++ src/components/navigation-bar/navigation-bar.jsx
@@ -1,40 +1,45 @@
 import { Navbar, Container, Nav } from "react-bootstrap";
 import { Link } from "react-router-dom";
 
-export const NavigationBar = ({ user, onLoggedOut }) => {
-  return (
-    <Navbar bg="light" expand="lg">
-      <Container>
-        <Navbar.Brand as={Link} to="/">
-          My Flix
-        </Navbar.Brand>
-        <Navbar.Toggle aria-controls="basic-navbar-nav" />
-        <Navbar.Collapse id="basic-navbar-nav">
-          <Nav className="me-auto">
-            
-            <Nav.Link className='nav-text font-style' href='/'>
-              Home
-            </Nav.Link>
+export const NavigationBar = ({ user, setUser, setToken }) => {
 
-            <Nav.Link className='nav-text font-style' href='/users'>
-              Profile
-            </Nav.Link>
+    if (!user) {
+        return null;
+    }
 
-            <Nav.Link
-              className='font-style'
-              href='#'
-              onClick={() => {
-                setUser(null);
-                setToken(null);
-                localStorage.clear();
-              }}
-              >
-              Logout
-            </Nav.Link>
+    return (
+        <Navbar bg="light" expand="lg">
+            <Container>
+                <Navbar.Brand as={Link} to="/">
+                    My Flix
+                </Navbar.Brand>
+                <Navbar.Toggle aria-controls="basic-navbar-nav" />
+                <Navbar.Collapse id="basic-navbar-nav">
+                    <Nav className="me-auto">
+                    
+                    <Nav.Link className='nav-text font-style' href='/'>
+                        Home
+                    </Nav.Link>
 
-          </Nav>
-        </Navbar.Collapse>
-      </Container>
-    </Navbar>
-  );
+                    <Nav.Link className='nav-text font-style' href='/users'>
+                        Profile
+                    </Nav.Link>
+
+                    <Nav.Link
+                        className='font-style'
+                        href='#'
+                        onClick={() => {
+                        setUser(null);
+                        setToken(null);
+                        localStorage.clear();
+                        }}
+                        >
+                        Logout
+                    </Nav.Link>
+
+                    </Nav>
+                </Navbar.Collapse>
+            </Container>
+        </Navbar>
+    );
 };
